refactor(js): make map an option in entry function
