+++ js/src/degausser.js
@@ -2,14 +2,12 @@ import { StringCollector } from './stringCollector'
 import { MapCollector } from './mapCollector'
 import { walkDOM } from './domWalker'
 
-export const degausser = (parentNode) => {
-  const collector = new StringCollector()
+export const degausser = (parentNode, options = {}) => {
+  let collector = new StringCollector()
 
-  return walkDOM(parentNode, collector)
-}
-
-export const mapWithDegausser = parentNode => {
-  const collector = new MapCollector()
+  if (options.map) {
+    collector = new MapCollector()
+  }
 
   return walkDOM(parentNode, collector)
 }
