+++ js/test/suites/links.test.js
@@ -8,7 +8,17 @@ describe(`Testing Links`, () => {
     test(`Testing ${element.i}`, () => {
       document.documentElement.innerHTML = element.i
 
-      expect(degausser(document.documentElement)).toBe(element.o)
+      const output = degausser(document.documentElement)
+      expect(output).toBe(element.o)
+
+      const map = degausser(document.documentElement, {map: true})
+      for (const mapSection of map) {
+        const sliced = output.slice(
+          mapSection.start,
+          mapSection.start + mapSection.length,
+        )
+        expect(sliced).toMatch(mapSection.content)
+      }
     })
   })
 })
